+++ atmega328p/atmega328p/usart.h
@@ -10,6 +10,6 @@
 #define USART_H_
 void init_usart(unsigned int baudrate);
 unsigned char usart_receive( void );
-void usart_transmit( uint8_t data );
+void usart_transmit( unsigned int data );
 void usart_transmit_string( char s[] );
 #endif /* USART_H_ */
